rtt_ros_integration: fix Header.h include
This hopefully fixes the issue completely now :).
The Header.h file used to be in roslib, but moved (since Diamondback) to
std_msgs. Although not in the std_msgs/include folder, the export flags are set
by rosbuild to use '#include <std_msgs/Header.h>'. Tested with Diamondback and
working with custom typekit generation.

Signed-off-by: Steven Bellens <steven.bellens@mech.kuleuven.be>