rtt_dynamic_reconfigure: fixed property updating from config in AutoConfig
The implementation of propertiesFromConfig() for rtt_dynamic_reconfigure::AutoConfig was using RTT::refreshProperties()
to update a PropertyBag from its internal PropertyBag, which is a no-op if the bag was empty before. The function is
called with an empty, temporary bag from Server::setConfigCallback() and therefore no reconfigure callback actually
changed the component's properties.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>