rtt_dynamic_reconfigure: added support for Property composition and decomposition and fixed reconfiguration of nested properties
Reconfiguration of nested properties (properties inside a PropertyBag) was broken because AutoConfig::__fromMessage__() aborted
too early if a property name was not found in the Config message, even if it is the name of a nested bag.

Composition support now allows to use the reconfigure service even for complex properties with complex types that support
composition (e.g. types from kdl_typekit or eigen_typekit).