rtt_dynamic_reconfigure: fixed potential deadlock in refresh() for RTT prior to 2.9
We do not know for sure which thread is calling this method/operation, but we can check if the current
thread is the same as the thread that will process the update/notify operation. If yes, we clone the
underlying OperationCaller implementation and set the caller to the processing engine. In this case
RTT <2.9 should always call the operation directly as if it would be a ClientThread operation:
https://github.com/orocos-toolchain/rtt/blob/toolchain-2.8/rtt/base/OperationCallerInterface.hpp#L79

RTT 2.9 and above already checks the caller thread internally and therefore does not require this hack.