typekits: move typekits back to https://github.com/orocos/rtt_common_msgs
This repository has been merged into rtt_ros_integration in https://github.com/orocos/rtt_ros_integration/pull/6.
While it is still true that some packages and tests in this repo depend on
typekit packages, they are very rarely updated and can be skipped for most
releases of rtt_ros_integration. This saves some time for doing the releases
and resources on the ROS build farm.

For source builds both repositories have to be added to the workspace because
of circular dependencies.