rtt_roscomm: replace add_file_dependencies() macro call by a custom target
Fixes build issues with parallel builds. If multiple targets depend on the same generated header file,
the file gets generated more than once, too (with CMake 3.5 in Ubuntu xenial).
Each invocation of the compiler might therefore see an inconsistent state of the generated header file
if another process associated with another target is about to overwrite the same file. The solution is
to replace the file-level dependencies by an explicit rtt-package-generate_boost_headers custom target
that all library targets depend on.

The dependencies on the current CMake list file (CMAKE_CURRENT_LIST_FILE) are not required because if
the file would change, cmake needs to run again and is normally executed automatically. If not, the
file dependency does not help either because the contents of the generated source files did not change.

Changes in upstream message or service definitions (.msg or .srv files) are already covered by the
declared dependencies of the generated boost headers or the implicit target dependency from including
the generated service headers.