rtt_rosparam: add a new RTT::DataSource that links to ROS parameters
The new RTT::DataSource keeps a reference name to a ROS parameter
so that .get() reads in the parameter and .set() sets out the
parameter value assigned to the data source.
The ROSParamService provides a new operation to add properties
based on ROS parameters. (for now only for Boolean types.)