rtt_roscomm: shortcut ROS spinner callback queue
The patch introduces a new "queue" without a queue to
bypass the global CallbackQueue of ROS for the subscriptions
of Orocos messages.
Instead, the network thread will immediately execute the
callback, and will pass it to the global CallbackQueue
only in case that it fails to execute with TryAgain status.
Helper functions are provided to generate a subscriber with
a custom implementation of CallbackQueueInterface.
This method is chosen for the subscribers of the message
transporter.